refactor: change the Layout, and add ErrorBoundary to the app
